Database Sharding/demo/requirements.txt
flask==1.1.2
psycopg==2-2.8.6
uhashring
jsonify